tweaking \`phantomsCapped\`
Previously, the PhantomSpawner mixin would use the exact same Info object that would be used in "normal" mob spawning, so if the mobcap filled up in the natural spawning phase, that would prevent any phantoms from spawning. Now, a deep copy of the Info class is made when its generated, and that is passed to the PhantomSpawner mixin, so even if the mob cap fills up during natural spawning, phantoms can still spawn if mobcap wasn't full beforehand.

i didn't need to write an essay but i wanted to lmao